requirements.txt
appdirs==1.4.3
appnope==0.1.0
APScheduler==3.6.3
attrs==19.3.0
authy==2.2.5
backcall==0.1.0
black==19.10b0
certifi==2020.4.5.1
chardet==3.0.4
click==7.1.1
decorator==4.4.2
Flask==1.1.2
flask-redis==0.4.0
gunicorn==20.0.4
idna==2.9
ipython==7.13.0
ipython-genutils==0.2.0
isort==4.3.21
itsdangerous==1.1.0
jedi==0.16.0
Jinja2==2.11.1
MarkupSafe==1.1.1
parso==0.6.2
pathspec==0.8.0
pexpect==4.8.0
pickleshare==0.7.5
prompt-toolkit==3.0.5
ptyprocess==0.6.0
Pygments==2.6.1
PyJWT==1.7.1
python-dotenv==0.12.0
pytz==2019.3
redis==3.4.1
regex==2020.4.4
requests==2.23.0
rq==1.3.0
simplejson==3.17.0
six==1.14.0
spotipy==2.10.0
toml==0.10.0
traitlets==4.3.3
twilio==6.38.0
typed-ast==1.4.1
tzlocal==2.0.0
urllib3==1.25.8
wcwidth==0.1.9
Werkzeug==1.0.1
